pyproject.toml
[project]
name = "f1_pipeline"
version = "0.1.0"
description = "Formula 1 analytics pipeline using dbt and PostgreSQL"
dependencies = [
    "dbt-postgres>=1.7.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.2",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "tenacity>=9.1.2"
]
requires-python = ">=3.10"

#[tool.setuptools]
#package-dir = {"" = "src_DELETE"}
#
#[tool.setuptools.packages.find]
#where = ["src_DELETE"]
